+++ previous-versions/banking/src/codatbanking/sdkconfiguration.py
@@ -20,8 +20,9 @@ class SDKConfiguration:
     server_idx: int = 0
     language: str = 'python'
     openapi_doc_version: str = '3.0.0'
-    sdk_version: str = '0.28.1'
-    gen_version: str = '2.131.1'
+    sdk_version: str = '0.29.0'
+    gen_version: str = '2.139.1'
+    user_agent: str = 'speakeasy-sdk/python 0.29.0 2.139.1 3.0.0 codat-banking'
     retry_config: RetryConfig = None
 
     def get_server_details(self) -> tuple[str, dict[str, str]]: