+++ previous-versions/banking/RELEASES.md
@@ -19,3 +19,13 @@ Based on:
 - [python v0.28.1] previous-versions/banking
 ### Releases
 - [PyPI v0.28.1] https://pypi.org/project/codat-banking/0.28.1 - previous-versions/banking
+
+## 2023-09-29 00:02:53
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Banking.yaml
+- Speakeasy CLI 1.91.3 (2.139.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.29.0] previous-versions/banking
+### Releases
+- [PyPI v0.29.0] https://pypi.org/project/codat-banking/0.29.0 - previous-versions/banking